It is Chef's birthday. You know that Chef's favourite number is X. You also know that Chef loves averages. Therefore you decide it's best to gift Chef 3 integers A1,A2,A3, such that:
The mean of A1,A2 and A3 is X.
1≤A1,A2,A3≤1000.
A1,A2 and A3 are distinct.
Output any suitable A1,A2 and A3 which you could gift to Chef.

As a reminder, the mean of three numbers P,Q,R is defined as: mean(P,Q,R)=P+Q+R3.

For example, mean(2,3,5)=2+3+53=103=3.333¯, mean(2,2,5)=2+2+53=93=3.